Update inner circuit id test
