+++ dpc/src/lib.rs
@@ -82,9 +82,6 @@ pub use errors::*;
 pub mod traits;
 pub use traits::*;
 
-#[cfg(test)]
-mod tests;
-
 pub mod prelude {
     pub use crate::{account::*, block::*, errors::*, traits::*};
 }
