+++ dpc/src/testnet2/tests.rs
@@ -15,16 +15,16 @@
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
 use crate::{
-    testnet1::{instantiated::Components, parameters::PublicParameters, Testnet1Components},
+    testnet2::{instantiated::Components, parameters::PublicParameters, Testnet2Components},
     traits::DPCComponents,
 };
 use snarkvm_algorithms::traits::{CRH, SNARK};
 use snarkvm_utilities::{to_bytes, ToBytes};
 
-fn testnet1_inner_circuit_id() -> anyhow::Result<Vec<u8>> {
+fn testnet2_inner_circuit_id() -> anyhow::Result<Vec<u8>> {
     let parameters = PublicParameters::<Components>::load(false)?;
 
-    let inner_snark_vk: <<Components as Testnet1Components>::InnerSNARK as SNARK>::VerifyingKey =
+    let inner_snark_vk: <<Components as Testnet2Components>::InnerSNARK as SNARK>::VerifyingKey =
         parameters.inner_snark_parameters.1.clone().into();
 
     let inner_circuit_id = <<Components as DPCComponents>::InnerCircuitIDCRH as CRH>::hash(
@@ -35,12 +35,14 @@ fn testnet1_inner_circuit_id() -> anyhow::Result<Vec<u8>> {
     Ok(to_bytes![inner_circuit_id]?)
 }
 
+/// TODO (howardwu): Update this to the correct inner circuit ID when the final parameters are set.
+#[ignore]
 #[test]
-fn test_inner_circuit_sanity_check() {
-    let expected_testnet1_inner_circuit_id = vec![
+fn test_testnet2_inner_circuit_sanity_check() {
+    let expected_testnet2_inner_circuit_id = vec![
         132, 243, 19, 234, 73, 219, 14, 105, 124, 12, 23, 229, 144, 168, 24, 163, 93, 33, 139, 247, 16, 201, 132, 0,
         141, 28, 29, 2, 131, 75, 18, 78, 248, 57, 118, 61, 81, 53, 11, 91, 196, 233, 80, 186, 167, 144, 163, 0,
     ];
-    let candidate_testnet1_inner_circuit_id = testnet1_inner_circuit_id().unwrap();
-    assert_eq!(expected_testnet1_inner_circuit_id, candidate_testnet1_inner_circuit_id);
+    let candidate_testnet2_inner_circuit_id = testnet2_inner_circuit_id().unwrap();
+    assert_eq!(expected_testnet2_inner_circuit_id, candidate_testnet2_inner_circuit_id);
 }
